# core/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Deck OS - JARVIS Cyberdeck Command Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /system/events:
    get:
      operationId: getSystemEvents
      tags: [system]
      summary: Get recent system events and alerts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Recent system events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemEventList"
...
components:
  schemas:
...
    SystemEventList:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/SystemEvent"
      required:
        - events
    SystemEvent:
      type: object
      properties:
        id:
          type: string
        level:
          type: string
          enum: [info, warning, error, critical]
        message:
          type: string
        source:
          type: string
        timestamp:
          type: string
          format: date-time
        data:
          type: object
          additionalProperties: true
          nullable: true
      required:
        - id
        - level
        - message
        - source
        - timestamp
...